[REF] rename linux engine files
